% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-fcast.R
\name{add_truth_data_ds_fcast}
\alias{add_truth_data_ds_fcast}
\title{Add truth data to forecasts}
\usage{
add_truth_data_ds_fcast(x, truth_data)
}
\arguments{
\item{x}{\code{\link[=ds_fcast]{ds_fcast()}}}

\item{truth_data}{states data}
}
\description{
Add truth data to forecasts
}
