% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demspaces-package.R
\docType{package}
\name{demspaces-package}
\alias{demspaces}
\alias{demspaces-package}
\title{demspaces: Functions for the Democratic Spaces Project}
\description{
Functions to support the Democratic Spaces project.
}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

}
\keyword{internal}
