% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demspacesR-package.R
\docType{package}
\name{demspacesR-package}
\alias{demspacesR}
\alias{demspacesR-package}
\title{demspacesR: Functions for the Democratic Spaces Project}
\description{
Functions to support the Democratic Spaces project.
}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

}
\keyword{internal}
