% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-fcast.R
\name{ds_fcast}
\alias{ds_fcast}
\title{ds_fcast constructor}
\usage{
ds_fcast(x)
}
\arguments{
\item{x}{a data frame of forecasts}
}
\description{
ds_fcast constructor
}
\examples{
data("states")

mdl <- ds_logistic_reg("v2x_veracc_osp", states)
preds <- predict(mdl, new_data = states)
head(ds_fcast(preds))
}
