% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg.R
\name{ds_logistic_reg}
\alias{ds_logistic_reg}
\title{Logistic regression}
\usage{
ds_logistic_reg(space, data, normalize = FALSE)
}
\arguments{
\item{space}{(\code{character(1)}) \cr
Name of the V-Dem indicator for a democratic space}

\item{data}{(\code{\link[base:data.frame]{base::data.frame()}})
A data frame or similar object}

\item{normalize}{Standardize data using \code{\link[=make_standardizer]{make_standardizer()}}?}
}
\description{
Demonstrates the interface. Otherwise only difference is that it internally
normalizes input data before fitting and predicting.
}
\examples{
data("states")

mdl <- ds_logistic_reg("v2x_veracc_osp", states)
preds <- predict(mdl, new_data = states)

}
\seealso{
Other Other DS models: 
\code{\link{ds_logistic_reg_featx}()},
\code{\link{ds_reg_logreg}()}
}
\concept{Other DS models}
\concept{ds_model}
