% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-logreg.R
\name{ds_reg_logreg}
\alias{ds_reg_logreg}
\title{Reguarized logistic regression}
\usage{
ds_reg_logreg(space, data, ...)
}
\arguments{
\item{space}{(\code{character(1)}) \cr
Name of the V-Dem indicator for a democratic space}

\item{data}{(\code{\link[base:data.frame]{base::data.frame()}})
A data frame or similar object}

\item{...}{document other arguments}
}
\description{
Demonstrates the interface. Otherwise only difference is that it internally
normalizes input data before fitting and predicting.
}
\examples{
data("states")

mdl <- ds_reg_logreg("v2x_veracc_osp", states, alpha_n = 1)
preds <- predict(mdl, new_data = states)

}
\seealso{
Other Other DS models: 
\code{\link{ds_logistic_reg_featx}()},
\code{\link{ds_logistic_reg}()}
}
\concept{Other DS models}
\concept{ds_model}
