% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{ds_rf}
\alias{ds_rf}
\title{Random forest}
\usage{
ds_rf(space, data, mtry = NULL, ...)
}
\arguments{
\item{space}{(\code{character(1)}) \cr
Name of the V-Dem indicator for a democratic space}

\item{data}{(\code{\link[base:data.frame]{base::data.frame()}})
A data frame or similar object}

\item{mtry}{see \code{\link[ranger:ranger]{ranger::ranger()}}; if NULL, this will be tuned via
out-of-bag error (see "ds_rf_tuner" in the source code in "rf.R").}

\item{...}{other arguments passed to \code{\link[ranger:ranger]{ranger::ranger()}}}
}
\description{
Demonstrates the interface. Otherwise only difference is that it internally
normalizes input data before fitting and predicting.
}
\examples{
data("states")

mdl <- ds_rf("v2x_veracc_osp", states)
preds <- predict(mdl, new_data = states)

}
