% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{ds_rf_tuner}
\alias{ds_rf_tuner}
\title{Wrapper to only tune mtry
Based on tunerf results}
\usage{
ds_rf_tuner(x, y, mtry, ...)
}
\description{
Wrapper to only tune mtry
Based on tunerf results
}
\keyword{internal}
