% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg.R
\name{logistic_reg}
\alias{logistic_reg}
\title{Logistic regression}
\usage{
logistic_reg(x, y)
}
\arguments{
\item{x}{Data frame with features.}

\item{y}{Binary vector indicating outcome event.}
}
\description{
Standardized interface for logistic regression
}
\examples{
library(modeldata)
data(credit_data)

mdl <- logistic_reg(credit_data[, setdiff(colnames(credit_data), "Status")],
                    credit_data$Status)

}
\seealso{
Other Other base models: 
\code{\link{logistic_reg_featx}()},
\code{\link{reg_logreg}()}
}
\concept{Other base models}
\concept{base_model}
