% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg-featx.R
\name{logistic_reg_featx}
\alias{logistic_reg_featx}
\title{Logistic regression with feature extraction}
\usage{
logistic_reg_featx(x, y, n_comp = 5)
}
\arguments{
\item{x}{Data frame with features.}

\item{y}{Binary vector indicating outcome event.}

\item{n_comp}{Number of components to keep.}
}
\description{
Standardized interface for logistic regression with a feature extraction
step, e.g. PCA.
}
\examples{
library(modeldata)
data(credit_data)
credit_data <- credit_data[complete.cases(credit_data), ]

mdl <- logistic_reg_featx(
  credit_data[, setdiff(colnames(credit_data), "Status")],
  credit_data$Status)

}
\seealso{
Other Other base models: 
\code{\link{logistic_reg}()},
\code{\link{reg_logreg}()}
}
\concept{Other base models}
\concept{base_model}
