% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg-featx.R
\name{new_ds_logistic_reg_featx}
\alias{new_ds_logistic_reg_featx}
\title{Constructor}
\usage{
new_ds_logistic_reg_featx(up_mdl, down_mdl, standardize, yname, n_comp)
}
\description{
Constructor
}
\keyword{internal}
