% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-logreg.R
\name{plot.reg_logreg}
\alias{plot.reg_logreg}
\title{Plot reg_logreg}
\usage{
\method{plot}{reg_logreg}(x, type = NULL, base = FALSE, ...)
}
\arguments{
\item{x}{see \code{\link[=reg_logreg]{reg_logreg()}}}

\item{type}{one of "grid", "lambda", "alpha"; see examples}

\item{base}{use base plot, not ggplot2?}

\item{...}{not used}
}
\description{
Plot reg_logreg
}
