% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{predict.ds_rf}
\alias{predict.ds_rf}
\title{Predict Method for DS random forest model}
\usage{
\method{predict}{ds_rf}(object, new_data, cutpoint = NULL, ...)
}
\arguments{
\item{object}{a fitted \code{\link[=ds_rf]{ds_rf()}} model}

\item{new_data}{predictor data}

\item{cutpoint}{optional cutpoint used to adjust forecasts, see details}

\item{...}{not used}
}
\description{
Create opening and closing predictions for a fitted \code{\link[=ds_rf]{ds_rf()}} model.
}
\details{
The "cutpoints" argument can optionally be used to adjust the raw
}
