% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{Random forest}
\usage{
rf(x, y, ...)
}
\arguments{
\item{x}{Data frame with features.}

\item{y}{Binary vector indicating outcome event.}

\item{...}{other arguments passed to \code{\link[ranger:ranger]{ranger::ranger()}}}
}
\description{
Standardized interface for ...
}
\examples{
library(modeldata)
data(credit_data)
credit_data <- credit_data[complete.cases(credit_data), ]

mdl <- rf(credit_data[, setdiff(colnames(credit_data), "Status")],
          credit_data$Status)

}
