% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-fcast.R
\name{score_ds_fcast}
\alias{score_ds_fcast}
\title{Score forecasts}
\usage{
score_ds_fcast(x, truth_data)
}
\arguments{
\item{x}{\code{\link[=ds_fcast]{ds_fcast()}}}

\item{truth_data}{states data}
}
\description{
Score forecasts
}
\examples{
data("states")

mdl <- ds_logistic_reg("v2x_veracc_osp", states)
preds <- predict(mdl, new_data = states)
fcasts <- ds_fcast(preds)
score_ds_fcast(fcasts, states)
}
