% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spaces}
\alias{spaces}
\title{Democratic spaces}
\source{
Own work
}
\value{
\item{spaces}{a data frame}
}
\description{
List of the democratic space concepts and indicator variables
}
\examples{
data("spaces")
spaces
}
\keyword{datasets}
