% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{States Data}
\source{
Own work
}
\value{
\item{states}{a data frame}
}
\description{
Example data for tests
}
\examples{
data("states")
str(states)
}
\keyword{datasets}
